more viewfactories simplified
