+++ widgets/src/commonMain/kotlin/dev/icerock/moko/widgets/factory/ContainerViewFactory.kt
@@ -1,0 +1,18 @@
+/*
+ * Copyright 2019 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package dev.icerock.moko.widgets.factory
+
+import dev.icerock.moko.widgets.ContainerWidget
+import dev.icerock.moko.widgets.core.ViewFactory
+import dev.icerock.moko.widgets.style.background.Background
+import dev.icerock.moko.widgets.style.view.MarginValues
+import dev.icerock.moko.widgets.style.view.PaddingValues
+import dev.icerock.moko.widgets.style.view.WidgetSize
+
+expect class ContainerViewFactory(
+    padding: PaddingValues? = null,
+    margins: MarginValues? = null,
+    background: Background? = null
+) : ViewFactory<ContainerWidget<out WidgetSize>>
